requirements.txt
#numpy==1.16.2
#scipy==1.2.0
FireWorks==1.8.7
pymatgen_diffusion==2019.2.28
monty==2.0.2
paramiko==2.4.2
tqdm==4.31.1
six==1.12.0
pybtex==0.22.2
ruamel.yaml==0.15.89
#pandas==0.23.4
#pymatgen==2019.03.27
custodian==2019.2.10
networkx==2.2
#pydash==4.7.3
